Split bugsigdb_analyzer.py into a package (1425 lines -> 7 focused files)
The largest, most architecturally tangled file in the repo mixed XML/PMC
parsing, text truncation, field-result building, JSON/heuristic parsing,
postprocessing, LLM-call orchestration, and two top-level async entry
points into one 1425-line module. Pure relocation, zero logic changes -
verified via AST-level diffing of every top-level def/class against the
original (git HEAD) content.

New layout:
  __init__.py          - module docstring (output contract) + re-exports
  singletons.py         - get_unified_qa/get_pubmed_retriever/get_cache_manager
  constants.py            - FIELD_KEYS, STATUS_COLUMNS, EXTRACTION_PROMPT
  parsing.py                - PMC XML section parsing, text truncation/prep
  field_extraction.py         - FieldDict builders, JSON/heuristic parsers,
                                  postprocessing, LLM extraction call
  simple_analysis.py            - analyze_paper_simple (v1 API)
  rag_analysis.py                  - analyze_paper_with_rag (v2 API),
                                       analyze_single_field (backward compat)

Splitting broke several @patch("app.services.bugsigdb_analyzer.X") test
targets (get_unified_qa, get_pubmed_retriever, get_cache_manager,
is_in_bugsigdb, PubMedRetriever, UnifiedQA, CacheManager) - patching the
module attribute no longer reached the binding actually used at call
time once the getters/imports lived in separate submodules. Fixed by
having each submodule do `import app.services.bugsigdb_analyzer as _pkg`
and call through `_pkg.X(...)` instead of a directly-imported name, so
the lookup happens at call time against the package's own (patchable)
namespace rather than being bound at import time. Verified directly:
every test that patches one of these names still passes, not just the
aggregate count.

Also fixed one credential-masking gap found while reviewing the split:
analyze_paper_simple's cache-store failure handler logged the raw,
unmasked exception (the equivalent line in analyze_paper_with_rag was
already fixed in an earlier commit this session; this one was missed).

Full 593-test suite passes unchanged; black/flake8 clean; mypy shows the
same 4 pre-existing errors as before (unrelated to this change, already
non-blocking per CI policy). Confirmed every router/test import site
(grepped project-wide) still resolves correctly, and smoke-tested
importing the full FastAPI app module end-to-end.
